requirements.txt
Pillow == 8.0.1
aeverall-testing-pypi == 0.1
aeveralltestingpypi == 0.1
ais-dom == 2021.6.7
ansar == 1.0.86
atacgraph == 1.0
copy-windows-10-lock-screen-wallpaper == 1.1.4
crawler == 0.0.2
gee == 5
homeassistant == 2021.6.6
iob == 0.0.2
ipython == 7.19.0
item == 0.0.3
keras_nightly == 2.5.0.dev2021032900
maizepagecontacts == 0.0.2
mediapipe == 0.8.4.2
node-enigma == 0.1.0
nomad-script == 1.082
nowyou-grafits-script-python == 1.1.3
opencv_python == 4.5.2.52
poc-t == 1.7.3
pytaraxa == 0.0.4
scunet == 1.1.0
spirent-testpack-framework == 1.0
spravka == 0.0.14
startpro == 1.2.9
tensorflow == 2.5.0
test-topsis == 1.0.0
tools == 0.1.9
txinterfacetest == 0.0.5
ucas-wifi == 1.0.1
video-konverter == 0.0.1
numpy == 1.19.2
